refactor: reorganize shared evaluation folders for semantics
