+++ src/front-end/typescript/lib/http/api/proposal/sprint-with-us/team-questions/consensus.ts
@@ -1,5 +1,5 @@
 import * as crud from "front-end/lib/http/crud";
-import * as Resource from "shared/lib/resources/question-evaluation/sprint-with-us";
+import * as Resource from "shared/lib/resources/evaluations/sprint-with-us/team-questions";
 import { Id } from "shared/lib/types";
 import { rawSWUTeamQuestionResponseEvaluationToSWUTeamQuestionResponseEvaluation } from "front-end/lib/http/api/question-evaluation/sprint-with-us";
