+++ src/front-end/typescript/lib/pages/opportunity/sprint-with-us/edit/tab/index.ts
@@ -23,7 +23,7 @@ import { User } from "shared/lib/resources/user";
 import { adt, Id } from "shared/lib/types";
 import { SWUProposalSlim } from "shared/lib/resources/proposal/sprint-with-us";
 import { GUIDE_AUDIENCE } from "front-end/lib/pages/guide/view";
-import { SWUTeamQuestionResponseEvaluation } from "shared/lib/resources/question-evaluation/sprint-with-us";
+import { SWUTeamQuestionResponseEvaluation } from "shared/lib/resources/evaluations/sprint-with-us/team-questions";
 
 // Parent page types & functions.
